add delete nd store function
